As skilltries are uint64_t since revision 6233, the cache also must be uint64_t.

+++ src/vocation.h
@@ -106,7 +106,7 @@ protected:
 	typedef std::map<uint32_t, uint64_t> manaCacheMap;
 	manaCacheMap cacheMana;
 
-	typedef std::map<uint32_t, uint32_t> skillCacheMap;
+	typedef std::map<uint32_t, uint64_t> skillCacheMap;
 	skillCacheMap cacheSkill[SKILL_LAST + 1];
 };
